Enhance the Cloud.ping method
The ping method use the limit argument in clients listing methods. This
argument allow to prevent of getting a large response from servers.

Fixes the bug #PH-81. The glance.images.list() returns a generator and
this generator retrieves images from the server only on the first
iteration. The first iteration of the generator should happen to produce
the request.