Add admin role for pumphouse main user
When migrating tenant, add pumphouse main user (admin) to the tenant in
destination cloud with 'admin' role. This is required on multiple occasions when
pumphouse needs to operate in tenant context without switching to another user.
Especially useful for a project workload migration as many resources only have
tenant as an owner.